// src/app/models/Discounts.ts
export interface IDiscountsAndAdditionalCoversMDL {
    IsNCBForward: boolean
    NCBPer: number
    NCBLevel: number
    IsAntiTheft: boolean
    IsAA: boolean
    IsHandicapped: boolean
    IsIMT23: boolean
    VoluntaryExcess: number
    DiscountPer: number
    Electrical: number
    NonElectrical: number
    BiFuel: number
}